Circle segment in timer uses calculated to values.
The timer now contains a circle segment that draws the segment up
to a point calculated by a function. This function, right now,
always returns [0,-1] (the center/top point of the circle), but we
can stub it in our tests.

The actual segment is a higher-order component created by a function.
This function gets a "circleSegment" function as a parameter, and now
I can pass the stub in our tests. The Slice does not decide whether
to draw the long or short side of the segment yet.

The <Timer/> should later decide to which percentage of the circle
the segment should be drawn, based on the current time, and pass
that as a prop to <Slice/> (at least that's my plan right now).