Refactoring: Moved functions to new file.
Moved functions percentFrom and segmentTo to their own file b/c
they are not directly realted to rendering the timer. Yes, together
they calculate a point on a circle from a given time. But they are
not directly related to th components that render the timer - they
can also be used on their own.