Simple segment for 10 mins drawn correctly.
The functions in segment.js now calculate percentage and the point
on the circle correclty. There was a wrong assumption in the Timer
regarding the order in which to draw the arc. Not it does the correct
order: Move to the left point on the circle, arc to the top point,
line to the center.