Config counts down correctly - but bad design.
OK, now the config components can start the timer and count down
correctly, but it does not pass the current time to the Timer view.
At least not the correct value.

That is hard to do for now because of the communication required.
Also, I am getting more and more dissatisfied with the design -
with having the time keeping inside the config.

I want to move stuff around a bit: Move config into the timer, move
the time keeping one level up - also into the timer. Before that, I
should probably move all the components that are insde Timer into
new files, to clean up a bit.