Started large(ish) refactoring: New config functionality.
The config now reports back the initial minutes and seconds immediately
after start *and* it reports back a running status when the start
button was pressed. All the other functionality is still there, but
now I want to move it to the parent (outer) component. With the start
time and the running status, I have (hopefully) all the infrastructure
I need to do that in place.